mapping changes and respose handling changes commit

+++ src/main/java/com/orgnisation/service/controllers/EmployeeController.java
@@ -3,6 +3,8 @@ package com.orgnisation.service.controllers;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -10,37 +12,61 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.orgnisation.service.model.Employee;
+import com.orgnisation.service.model.Response;
 import com.orgnisation.service.repository.EmployeeRepository;
 
 @RestController
+@RequestMapping("/employees")
 public class EmployeeController {
 
 	@Autowired
 	private EmployeeRepository employeeRepository;
-
-	@RequestMapping("/employees")
-	public List<Employee> getAllEmployees() {
-		return employeeRepository.findAll();
-
+	
+	@RequestMapping("/allEmployees")
+	public ResponseEntity<?> getAllEmployees() {
+		List<Employee> employees = employeeRepository.findAll();
+		if(!employees.isEmpty()) {
+			return new ResponseEntity<>(new Response(employees, "all employees"), HttpStatus.OK);
+		} else {
+			return new ResponseEntity<>(new Response("no employees for the request"), HttpStatus.INTERNAL_SERVER_ERROR);
+		}
 	}
 
-	@RequestMapping(method = RequestMethod.POST, value = "/employees/addEmployee")
+	@RequestMapping(method = RequestMethod.POST, value = "/addEmployee")
 	public void addEmployee(@RequestBody Employee employee) {
-		employeeRepository.save(employee);
+		if (employee != null)
+			employeeRepository.save(employee);
 	}
 
-	@RequestMapping("/employees/getEmployeeByDesignation/{designation}")
-	public List<Employee> getEmployeeByDesignation(@PathVariable String designation) {
-		return employeeRepository.getEmployeeByDesignation(designation);
+	@RequestMapping("/getEmployeeByDesignation/{designation}")
+	public ResponseEntity<?> getEmployeeByDesignation(@PathVariable String designation) {
+		List<Employee> employees = employeeRepository.getEmployeeByDesignation(designation);
+		if(!employees.isEmpty()) {
+			return new ResponseEntity<>(new Response(employees, "list of employees with designation : "+designation), HttpStatus.OK);
+		} else {
+			return new ResponseEntity<>(new Response("no employees with designation : "+designation), HttpStatus.INTERNAL_SERVER_ERROR);
+		}
 	}
 
-	@RequestMapping("/employees/getEmployeeByDomain/{domain}")
-	public List<Employee> getEmployeeByDomain(@PathVariable String domain) {
-		return employeeRepository.getEmployeeByDomain(domain);
+	@RequestMapping("/getEmployeeByDomain/{domain}")
+	public ResponseEntity<?> getEmployeeByDomain(@PathVariable String domain) {
+		List<Employee> employees = employeeRepository.getEmployeeByDomain(domain);
+		if(!employees.isEmpty()) {
+			return new ResponseEntity<>(new Response(employees, "list of employees with domain : "+domain), HttpStatus.OK);
+		} else {
+			return new ResponseEntity<>(new Response("no employees with domain : "+domain), HttpStatus.INTERNAL_SERVER_ERROR);
+		}
 	}
 
-	@RequestMapping(method = RequestMethod.DELETE, value = "/employees/deleteEmployeeByEmployeeName/{employeeName}")
+	@RequestMapping(method = RequestMethod.DELETE, value = "/deleteEmployee/{employeeName}")
 	public void deleteEmployeeByEmployeeName(@PathVariable String employeeName) {
-		employeeRepository.deleteEmployeeByEmployeeName(employeeName);
+		if (employeeName != null)
+			employeeRepository.deleteEmployeeByEmployeeName(employeeName);
+	}
+
+	@RequestMapping(method = RequestMethod.PUT)
+	public void updateEmployee(@RequestBody Employee employee) {
+		if (employee != null)
+			employeeRepository.save(employee);
 	}
 }
